Fix C++ compiler warnings in bytetrack sources
- Lapjv.cpp: silence unused parameter 'y' in _find_dense; fix signed/
  unsigned comparisons by casting vector::size() to int in all lapjv()
  index loops
- Utils.cpp: fix range-loop copy of at::Tensor to const reference
- BYTETracker.cpp: remove unused local variables (lost_frame_count,
  frame_id, last_frame_id); fix signed/unsigned comparison against
  config_.num_frames_to_keep_lost_tracks via static_cast<int64_t>
- HmTracker.cpp: fix signed/unsigned comparison in tensor size assert

+++ hockeymom/csrc/bytetrack/HmTracker.cpp
@@ -99,7 +99,7 @@ std::unordered_map<std::string, at::Tensor> HmTracker::track(
       } else if (item.first == kFrameId) {
         continue;
       }
-      assert(tensor.size(0) == id_count);
+      assert(tensor.size(0) == static_cast<int64_t>(id_count));
       // std::cout << item.first << " -> " << tensor.device()
       //           << std::endl;
       tensor = bool_mask_select(tensor, all_active_mask);
